fix referential syntax of updates and messages
