+++ src/components/Home/Home.js
@@ -5,7 +5,7 @@ import React, {
 // import { Link } from 'react-router-dom';
 // import { useSelector, useDispatch } from 'react-redux';
 import Trade from '../Trade/Trade.js';
-import Updates from '../Updates/Updates.js';
+import Updates from '../Updates/Messages.js';
 import ToggleBot from '../ToggleBot/ToggleBot'
 import TradeList from '../TradeList/TradeList'
 import Status from '../Status/Status'
